eliminate wrapper.tsx
