+++ packages/insomnia/src/ui/containers/app.tsx
@@ -1,17 +1,68 @@
-import React, { useEffect, useState } from 'react';
-import {  useSelector } from 'react-redux';
+import React, { FC, Fragment, lazy, Suspense, useEffect, useState } from 'react';
+import { useSelector } from 'react-redux';
+import { Route, Routes, useNavigate } from 'react-router-dom';
 
+import {
+  ACTIVITY_DEBUG,
+  ACTIVITY_HOME,
+  ACTIVITY_SPEC,
+  ACTIVITY_UNIT_TEST,
+} from '../../common/constants';
 import { database as db } from '../../common/database';
 import * as models from '../../models';
 import { ErrorBoundary } from '../components/error-boundary';
+import { AddKeyCombinationModal } from '../components/modals/add-key-combination-modal';
+import { AlertModal } from '../components/modals/alert-modal';
+import { AnalyticsModal } from '../components/modals/analytics-modal';
+import { AskModal } from '../components/modals/ask-modal';
+import { CodePromptModal } from '../components/modals/code-prompt-modal';
+import { CookieModifyModal } from '../components/modals/cookie-modify-modal';
+import { CookiesModalFC } from '../components/modals/cookies-modal';
+import { EnvironmentEditModal } from '../components/modals/environment-edit-modal';
+import { ErrorModal } from '../components/modals/error-modal';
+import { ExportRequestsModal } from '../components/modals/export-requests-modal';
+import { FilterHelpModal } from '../components/modals/filter-help-modal';
+import { GenerateCodeModal } from '../components/modals/generate-code-modal';
+import { GenerateConfigModal } from '../components/modals/generate-config-modal';
+import { GitBranchesModal } from '../components/modals/git-branches-modal';
+import { GitLogModal } from '../components/modals/git-log-modal';
+import { GitRepositorySettingsModal } from '../components/modals/git-repository-settings-modal';
+import { GitStagingModal } from '../components/modals/git-staging-modal';
+import { registerModal } from '../components/modals/index';
+import { LoginModal } from '../components/modals/login-modal';
+import { NunjucksModal } from '../components/modals/nunjucks-modal';
+import ProjectSettingsModal from '../components/modals/project-settings-modal';
+import { PromptModal } from '../components/modals/prompt-modal';
+import ProtoFilesModal from '../components/modals/proto-files-modal';
+import { RequestGroupSettingsModal } from '../components/modals/request-group-settings-modal';
+import { RequestRenderErrorModal } from '../components/modals/request-render-error-modal';
+import { RequestSettingsModal } from '../components/modals/request-settings-modal';
+import { RequestSwitcherModal } from '../components/modals/request-switcher-modal';
+import { ResponseDebugModal } from '../components/modals/response-debug-modal';
+import { SelectModal } from '../components/modals/select-modal';
+import { SettingsModal } from '../components/modals/settings-modal';
+import { SyncBranchesModal } from '../components/modals/sync-branches-modal';
+import { SyncDeleteModal } from '../components/modals/sync-delete-modal';
+import { SyncHistoryModal } from '../components/modals/sync-history-modal';
+import { SyncMergeModal } from '../components/modals/sync-merge-modal';
+import { SyncStagingModal } from '../components/modals/sync-staging-modal';
+import { WorkspaceDuplicateModal } from '../components/modals/workspace-duplicate-modal';
+import { WorkspaceEnvironmentsEditModal } from '../components/modals/workspace-environments-edit-modal';
+import { WorkspaceSettingsModal } from '../components/modals/workspace-settings-modal';
+import { WrapperModal } from '../components/modals/wrapper-modal';
 import { Toast } from '../components/toast';
-import { Wrapper } from '../components/wrapper';
 import withDragDropContext from '../context/app/drag-drop-context';
-import { GrpcProvider } from '../context/grpc';
+import { GrpcDispatchModalWrapper, GrpcProvider } from '../context/grpc';
 import { NunjucksEnabledProvider } from '../context/nunjucks/nunjucks-enabled-context';
+import { useGitVCS } from '../hooks/use-git-vcs';
+import { useVCS } from '../hooks/use-vcs';
 import {
+  selectActiveActivity,
   selectActiveApiSpec,
   selectActiveCookieJar,
+  selectActiveEnvironment,
+  selectActiveGitRepository,
+  selectActiveProject,
   selectActiveWorkspace,
   selectActiveWorkspaceMeta,
   selectEnvironments,
@@ -19,6 +70,65 @@ import {
   selectIsLoggedIn,
 } from '../redux/selectors';
 import { AppHooks } from './app-hooks';
+
+const lazyWithPreload = <T extends FC<any>>(
+  importFn: () => Promise<{ default: T }>
+) => {
+  const LazyComponent = lazy(importFn);
+  const preload = () => importFn();
+
+  return [LazyComponent, preload] as const;
+};
+
+const [WrapperHome, preloadWrapperHome] = lazyWithPreload(
+  () => import('../components/wrapper-home')
+);
+const [WrapperDebug, preloadWrapperDebug] = lazyWithPreload(
+  () => import('../components/wrapper-debug')
+);
+const [WrapperDesign, preloadWrapperDesign] = lazyWithPreload(
+  () => import('../components/wrapper-design')
+);
+const [WrapperUnitTest, preloadWrapperUnitTest] = lazyWithPreload(
+  () => import('../components/wrapper-unit-test')
+);
+
+preloadWrapperHome();
+preloadWrapperDebug();
+preloadWrapperDesign();
+preloadWrapperUnitTest();
+
+const LoadingIndicator = () => (<div
+  id="app-loading-indicator"
+  style={{
+    position: 'fixed',
+    top: '0',
+    left: '0',
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    height: '100%',
+    width: '100%',
+  }}
+>
+  <img src="./ui/images/insomnia-logo.svg" alt="Insomnia" />
+</div>);
+
+const ActivityRouter = () => {
+  const selectedActivity = useSelector(selectActiveActivity);
+  const activeWorkspace = useSelector(selectActiveWorkspace);
+  // If there is no active workspace, we want to navigate to home no matter what the previous activity was
+  const activity = activeWorkspace ? selectedActivity : ACTIVITY_HOME;
+  const navigate = useNavigate();
+
+  React.useEffect(() => {
+    if (activity) {
+      navigate(activity);
+    }
+  }, [activity, navigate]);
+
+  return null;
+};
 
 interface State {
   isMigratingChildren: boolean;
@@ -36,7 +146,25 @@ const App = () => {
   const environments = useSelector(selectEnvironments);
   const isLoggedIn = useSelector(selectIsLoggedIn);
   const isFinishedBooting = useSelector(selectIsFinishedBooting);
+  const activeGitRepository = useSelector(selectActiveGitRepository);
+  const activeEnvironment = useSelector(selectActiveEnvironment);
+  const activeProject = useSelector(selectActiveProject);
 
+  const vcs = useVCS({
+    workspaceId: activeWorkspace?._id,
+  });
+
+  const gitVCS = useGitVCS({
+    workspaceId: activeWorkspace?._id,
+    projectId: activeProject?._id,
+    gitRepository: activeGitRepository,
+  });
+
+  async function handleSetActiveEnvironment(activeEnvironmentId: string | null) {
+    if (activeWorkspaceMeta) {
+      await models.workspaceMeta.update(activeWorkspaceMeta, { activeEnvironmentId });
+    }
+  }
   // Ensure Children: Make sure cookies, env, and meta models are created under this workspace
   useEffect(() => {
     if (!activeWorkspace) {
@@ -81,7 +209,144 @@ const App = () => {
         <AppHooks />
         <div className="app" key={uniquenessKey}>
           <ErrorBoundary showAlert>
-            <Wrapper />
+            <div key="modals" className="modals">
+              <ErrorBoundary showAlert>
+                <AnalyticsModal />
+                <AlertModal ref={registerModal} />
+                <ErrorModal ref={registerModal} />
+                <PromptModal ref={registerModal} />
+                <WrapperModal ref={registerModal} />
+                <LoginModal ref={registerModal} />
+                <AskModal ref={registerModal} />
+                <SelectModal ref={registerModal} />
+                <FilterHelpModal ref={registerModal} />
+                <RequestRenderErrorModal ref={registerModal} />
+                <GenerateConfigModal ref={registerModal} />
+                <ProjectSettingsModal ref={registerModal} />
+                <WorkspaceDuplicateModal ref={registerModal} vcs={vcs || undefined} />
+                <CodePromptModal ref={registerModal} />
+                <RequestSettingsModal ref={instance => registerModal(instance, 'RequestSettingsModal')} />
+                <RequestGroupSettingsModal ref={registerModal} />
+
+                {activeWorkspace ? <>
+                  {/* TODO: Figure out why cookieJar is sometimes null */}
+                  {activeCookieJar ? <>
+                    <CookiesModalFC
+                      ref={registerModal}
+                    />
+                    <CookieModifyModal ref={registerModal} />
+                  </> : null}
+
+                  <NunjucksModal
+                    ref={registerModal}
+                    workspace={activeWorkspace}
+                  />
+
+                  {activeApiSpec ? <WorkspaceSettingsModal
+                    ref={registerModal}
+                    workspace={activeWorkspace}
+                    apiSpec={activeApiSpec}
+                  /> : null}
+                </> : null}
+
+                <GenerateCodeModal
+                  ref={registerModal}
+                  environmentId={activeEnvironment ? activeEnvironment._id : 'n/a'}
+                />
+
+                <SettingsModal ref={instance => registerModal(instance, 'SettingsModal')} />
+                <ResponseDebugModal ref={instance => registerModal(instance, 'ResponseDebugModal')} />
+
+                <RequestSwitcherModal ref={instance => registerModal(instance, 'RequestSwitcherModal')} />
+
+                <EnvironmentEditModal
+                  ref={registerModal}
+                  onChange={models.requestGroup.update}
+                />
+
+                <GitRepositorySettingsModal ref={registerModal} />
+
+                {activeWorkspace && gitVCS ? (
+                  <Fragment>
+                    <GitStagingModal ref={registerModal} workspace={activeWorkspace} vcs={gitVCS} gitRepository={activeGitRepository} />
+                    <GitLogModal ref={registerModal} vcs={gitVCS} />
+                    {activeGitRepository !== null && (
+                      <GitBranchesModal
+                        ref={registerModal}
+                        vcs={gitVCS}
+                        gitRepository={activeGitRepository}
+                      />
+                    )}
+                  </Fragment>
+                ) : null}
+
+                {activeWorkspace && vcs ? (
+                  <Fragment>
+                    <SyncStagingModal ref={registerModal} vcs={vcs} />
+                    <SyncMergeModal ref={registerModal} vcs={vcs} />
+                    <SyncBranchesModal ref={registerModal} vcs={vcs} />
+                    <SyncDeleteModal ref={registerModal} vcs={vcs} />
+                    <SyncHistoryModal ref={registerModal} vcs={vcs} />
+                  </Fragment>
+                ) : null}
+
+                <WorkspaceEnvironmentsEditModal
+                  ref={registerModal}
+                  handleSetActiveEnvironment={handleSetActiveEnvironment}
+                  activeEnvironmentId={activeEnvironment ? activeEnvironment._id : null}
+                />
+
+                <AddKeyCombinationModal ref={instance => registerModal(instance, 'AddKeyCombinationModal')} />
+                <ExportRequestsModal ref={registerModal} />
+
+                <GrpcDispatchModalWrapper>
+                  {dispatch => (
+                    <ProtoFilesModal
+                      ref={registerModal}
+                      grpcDispatch={dispatch}
+                    />
+                  )}
+                </GrpcDispatchModalWrapper>
+              </ErrorBoundary>
+            </div>
+
+            <Routes>
+              <Route
+                path="*"
+                element={
+                  <Suspense fallback={<LoadingIndicator />}>
+                    <WrapperHome
+                      vcs={vcs}
+                    />
+                  </Suspense>
+                }
+              />
+              <Route
+                path={ACTIVITY_UNIT_TEST}
+                element={
+                  <Suspense fallback={<LoadingIndicator />}>
+                    <WrapperUnitTest />
+                  </Suspense>
+                }
+              />
+              <Route
+                path={ACTIVITY_SPEC}
+                element={
+                  <Suspense fallback={<LoadingIndicator />}>
+                    <WrapperDesign />
+                  </Suspense>
+                }
+              />
+              <Route
+                path={ACTIVITY_DEBUG}
+                element={
+                  <Suspense fallback={<LoadingIndicator />}>
+                    <WrapperDebug />
+                  </Suspense>
+                }
+              />
+            </Routes>
+            <ActivityRouter />
           </ErrorBoundary>
 
           <ErrorBoundary showAlert>
